% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyStartR.R
\docType{package}
\name{ShinyStartR-package}
\alias{ShinyStartR}
\alias{ShinyStartR-package}
\title{ShinyStartR: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
Use four spaces when indenting paragraphs within the Description.
}
\details{
See \code{vignette('ShinyStartR')} for an overview.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kiwiroy/ShinyStartR}
  \item Report bugs at \url{https://github.com/kiwiroy/ShinyStartR/issues}
}

}
\author{
\strong{Maintainer}: Roy Storey \email{kiwiroy@users.noreply.github.com}

}
